# node/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: IPFS REST API

host: localhost:10011

basePath: /ipfs

schemes:
  - http
  - https

consumes:
  - application/json

produces:
  - application/json

paths:

###############################
########### MISC ##############
###############################

  /healthcheck:
    x-swagger-router-controller: misc.controller
    get:
      description: Check whether the IPFS platform is running
      operationId: healthcheck
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/genericObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"

            
####################################
####### DOCUMENTS MANAGEMENT #######
####################################

  /documents:
    x-swagger-router-controller: document.controller
    get:
      description: Get the document from a particular IPFS hash
      operationId: getData
      parameters:
        - name: hash
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
    post:
      description: Add documents to IPFS
      operationId: insertDocument
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: newDocument
          type: file
          description: The data to upload.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/genericObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"


###############################
########## DEFINITIONS ########
###############################          
            
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:


###############################
########### ARGUMENTS #########
############################### 
        

###############################
########## RESPONSES ##########
###############################  

  genericObjectResponse:
    type: object

      
###############################
############ ERRORS ###########
###############################        
    
  errorResponse:
    required:
      - message
    properties:
      message:
        type: string